Backend/requirements.txt
numpy
Pillow
scipy
scikit-image
matplotlib
flask
requests
tqdm